requirements.txt
numpy
opencv-python
opencv-contrib-python
uvicorn
boto3 
uuid
django-storages
djangorestframework
python-dotenv
gunicorn
psycopg2-binary
python-docx 
Pillow
whitenoise
dj-database-url
python-decouple    
whitenoise
weasyprint
cairocffi
pango
gobject